plant_recognition/backend/requirements.txt
# Plant Recognition Backend Dependencies

# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0

# LLM and RAG dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.20.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
opencv-python>=4.5.0
albumentations>=1.1.0

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Development
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0 